Introduced FileLock and number of jobs flag for better cluster interop
It turns out that joblib doesn't really work when running across
multiple nodes - which makes sense since that's quite a complex problem.
It seems we're capped at 39 jobs per node.

The solution is to just run multiple `bsub`s, each with 38
jobs/instances. The problem is when they all want to write to the result
file at the same time. Thus we introduce a file-locking mechanism that
creates a lock file which the processes have to acquire. Let's hope that
it works ;)