Add data loader and data module
When working with graph data, the data representation is actually a bit
complex. This is handled by the `MilpBipartiteData` class, which is
loaded by the `MilpDataset`. Additionally, the big csv is read and the
"simulation" results and configuration settings are also loaded.

Note that I did a benchmark to compare loading Milp representations by
calling `ecole.env.reset()` and pre-computing them and storing them as a
pickle file. It seems that the pickle approach is about 100x faster to
load, so I converted all the instances to pikle files.

Closes #5