Add mps_to_pkl file - rune once before training!
When training the data, loading the instances from the mps
representation takes quite a long time. I decided to just encode them
once as the MILP observation and then save them as pickle files. This
seems to be much faster and the training runs very well with this.

So, before running the training, convert your instances once by running
this script.