Make evaluation callback
The callback judges "how well we're doing" in the sense of the
competition.
It queries the NN for the best configuration choice and then computes
the mean primal_dual_integral given the data with that configuration
choice. Then it computes the percentile of that integral and returns the
average of this metric over all instances.